module3/requirements.txt
numpy==1.22.1
paho-mqtt==1.6.1